+++ src/index.ts
@@ -1,2 +1,4 @@
-import { pact } from "./eden-query";
-export { pact };
+import { createClient } from "./create-client";
+import { createUtils } from "./create-utils";
+import { pact } from "./legacy/eden-query";
+export { pact, createClient };